trabajando en clasificación de proyectos e interfaz de usuario

+++ Proyectogestionhoras/Proyectogestionhoras/Models/DTO/HorasSociosActDTO.cs
@@ -3,11 +3,11 @@
     public class HorasSociosActDTO
     {
         public string? NombreSocio { get; set; }
-        public Dictionary<string,int>? HorasPorActividad { get; set; }
+        public Dictionary<string,decimal>? HorasPorActividad { get; set; }
 
         public HorasSociosActDTO() { 
         
-            HorasPorActividad = new Dictionary<string,int>();
+            HorasPorActividad = new Dictionary<string,decimal>();
         }
     }
 }
